requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blis==0.7.11
CacheControl==0.14.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cryptography==45.0.3
cymem==2.0.11
es-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/es_core_news_lg-3.5.0/es_core_news_lg-3.5.0-py3-none-any.whl#sha256=44c60195ba508dc75b209f1e3328d12e38851e6e17705c7582bf7e8442f53529
fastapi==0.95.2
firebase-admin==6.1.0
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
logging==0.4.9.6
lxml==5.4.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
msgpack==1.1.0
murmurhash==1.0.13
numpy==1.23.5
packaging==25.0
pathlib==1.0.1
pathlib_abc==0.1.1
pathy==0.11.0
pillow==11.2.1
preshed==3.0.10
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==1.10.22
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-pptx==1.0.2
requests==2.32.3
rsa==4.9.1
smart-open==6.4.0
sniffio==1.3.1
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.27.0
thinc==8.1.12
tqdm==4.67.1
typer==0.7.0
typing-inspection==0.4.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.22.0
wasabi==1.1.3
weasel==0.4.1
XlsxWriter==3.2.3
